pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Daniil-Solo <daniil.solo1723@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.3.0"}
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
sqlalchemy = "^2.0.21"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
